Use Subscriber component correctly, temporarily remove histogram, revert VitessceGrid refactor
