+++ src/app/index.js
@@ -1,8 +1,6 @@
 import { createApp, Vitessce } from './app';
-import VitessceGrid from './VitessceGrid';
 
 export {
   createApp,
   Vitessce,
-  VitessceGrid,
 };
