+++ src/components/vega/index.js
@@ -1,3 +1,3 @@
 export { default as VegaPlot } from './VegaPlot';
-export { default as SetSizePlotSubscriber } from './SetSizePlotSubscriber';
-export { default as LayerHistogramSubscriber } from './LayerHistogramSubscriber';
+export { default as CellSetSizePlot } from './CellSetSizePlot';
+export { default as CellSetSizePlotSubscriber } from './CellSetSizePlotSubscriber';
