+++ src/components/vega/utils.js
@@ -5,6 +5,11 @@ import * as Vega from 'vega';
 import * as VegaLite from 'vega-lite';
 import { GRID_RESIZE } from '../../events';
 
+/**
+ * Custom hook
+ * @returns {array} `[width, height, containerRef]` where width and height
+ * are numbers and containerRef is a React ref.
+ */
 export function useGridItemSize() {
   const containerRef = useRef();
 
@@ -30,6 +35,10 @@ export function useGridItemSize() {
   return [width, height, containerRef];
 }
 
+/**
+ * Vega-Lite themes that can be passed to the `config` property
+ * of the vega-lite spec.
+ */
 export const VEGA_THEMES = {
   dark: {
     // The vega-themes dark theme.
@@ -59,6 +68,11 @@ export const VEGA_THEMES = {
   },
 };
 
+/**
+ * Get a `vl` Vega-Lite API object.
+ * @returns {object} The Vega-Lite API object, configured
+ * by registering the Vega and Vega-Lite libraries.
+ */
 export function createVegaLiteApi() {
   vl.register(Vega, VegaLite);
   return vl;
